+++ src/nvim/memline.c
@@ -69,6 +69,7 @@
 #include "nvim/spell.h"
 #include "nvim/strings.h"
 #include "nvim/term.h"
+#include "nvim/tempfile.h"
 #include "nvim/ui.h"
 #include "nvim/undo.h"
 #include "nvim/window.h"
