+++ src/nvim/misc1.c
@@ -52,6 +52,7 @@
 #include "nvim/strings.h"
 #include "nvim/tag.h"
 #include "nvim/term.h"
+#include "nvim/tempfile.h"
 #include "nvim/ui.h"
 #include "nvim/undo.h"
 #include "nvim/window.h"
