+++ src/nvim/quickfix.c
@@ -47,6 +47,7 @@
 #include "nvim/search.h"
 #include "nvim/strings.h"
 #include "nvim/term.h"
+#include "nvim/tempfile.h"
 #include "nvim/ui.h"
 #include "nvim/window.h"
 #include "nvim/os/os.h"
