+++ src/nvim/tempfile.h
@@ -1,0 +1,8 @@
+#ifndef NVIM_TEMPFILE_H
+#define NVIM_TEMPFILE_H
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "tempfile.h.generated.h"
+#endif
+
+#endif  // NVIM_TEMPFILE_H
